=cls Split sharding classes into separate files
